cleanup logger.library
